backend/requirements.txt
# DrugTox-AI Clean Backend Requirements
flask>=2.3.0
flask-cors>=4.0.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.2.0

# AI and Database Integration
groq>=0.4.0
supabase>=2.0.0
python-dotenv>=1.0.0